broke up util.py into smaller, more descriptive modules
