+++ text_extensions_for_pandas/cleaning/__init__.py
@@ -1,0 +1,29 @@
+#
+#  Copyright (c) 2020 IBM Corp.
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#  http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+#
+
+################################################################################
+# io module
+#
+# Functions in text_extensions_for_pandas that create DataFrames and convert
+# them to other formats.
+
+# Expose the public APIs that users should get from importing the top-level
+# library.
+
+from text_extensions_for_pandas.cleaning import train
+from text_extensions_for_pandas.cleaning import analysis
+from text_extensions_for_pandas.cleaning import preprocess
+
+__all__ = ['train', 'analysis', 'preprocess']
